Refataração do codigo é implementação do metodo para somar o total do pedido.

+++ spring-boot/src/main/java/br/com/vendas/dto/pedido/InformacaoIntemPedido.java
@@ -6,7 +6,7 @@ public class InformacaoIntemPedido {
 
     private String descricaoProduto;
 
-    private BigDecimal precoUnitario;
+    private Double precoUnitario;
 
     private Integer quantidade;
 
@@ -18,11 +18,11 @@ public class InformacaoIntemPedido {
         this.descricaoProduto = descricaoProduto;
     }
 
-    public BigDecimal getPrecoUnitario() {
+    public Double getPrecoUnitario() {
         return precoUnitario;
     }
 
-    public void setPrecoUnitario(BigDecimal precoUnitario) {
+    public void setPrecoUnitario(Double precoUnitario) {
         this.precoUnitario = precoUnitario;
     }
