+++ spring-boot/src/main/java/br/com/vendas/dto/produto/ProdutoRequestDTO.java
@@ -1,10 +1,8 @@
 package br.com.vendas.dto.produto;
 
 import javax.persistence.Column;
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 
 public class ProdutoRequestDTO {
 
@@ -13,7 +11,7 @@ public class ProdutoRequestDTO {
 
     @NotNull(message = "Preço obrigatorio")
     @Column(name = "preco_unitario")
-    private BigDecimal preco;
+    private Double preco;
 
     public String getDescricao() {
         return descricao;
@@ -23,11 +21,11 @@ public class ProdutoRequestDTO {
         this.descricao = descricao;
     }
 
-    public BigDecimal getPreco() {
+    public Double getPreco() {
         return preco;
     }
 
-    public void setPreco(BigDecimal preco) {
+    public void setPreco(Double preco) {
         this.preco = preco;
     }
 }
